ASA:ais getting closer
