+++ .pre-commit-config.yaml
@@ -1,45 +1,52 @@
 repos:
-    -
-        repo: https://github.com/pre-commit/mirrors-isort
-        rev: v5.10.1
-        hooks:
-            - id: isort
-              language_version: python
-              args: ["--profile", "black"]
-    -
-        repo: 'https://github.com/psf/black'
-        rev: 24.10.0
-        hooks:
-            - id: black
-              args: ['--safe']
-              language_version: python
-    -
-        repo: https://github.com/PyCQA/flake8
-        rev: 7.1.1
-        hooks:
-            - id: flake8
-              language_version: python
-              args: [
-                  # E501 let black handle all line length decisions
-                  # W503 black conflicts with "line break before operator" rule
-                  # E203 black conflicts with "whitespace before ':'" rule
-                  '--ignore=E501,W503,E203']
-    -
-        repo: 'https://github.com/chewse/pre-commit-mirrors-pydocstyle'
-        # 2.1.1
-        rev: v2.1.1
-        hooks:
-            - id: pydocstyle
-              language_version: python
-              args: [
-                 # Check for docstring presence only
-                 '--select=D1',
-                 # Don't require docstrings for tests
-                 '--match=(?!test).*\.py']
+  - repo: https://github.com/pre-commit/mirrors-isort
+    rev: v5.10.1
+    hooks:
+      - id: isort
+        language_version: python
+        args: ["--profile", "black"]
 
-    -
-        repo: https://github.com/pre-commit/mirrors-mypy
-        rev: 'v1.11.2'
-        hooks:
-            - id: mypy
-              args: ['--install-types', '--non-interactive', '--no-strict-optional', '--ignore-missing-imports']
+  - repo: 'https://github.com/psf/black'
+    rev: 24.10.0
+    hooks:
+      - id: black
+        args: ['--safe']
+        language_version: python
+
+  - repo: https://github.com/PyCQA/flake8
+    rev: 7.1.1
+    hooks:
+      - id: flake8
+        language_version: python
+        args: [
+            '--ignore=E501,W503,E203'
+          ]
+
+  # Remove the old pydocstyle repo configuration
+  # -
+  #     repo: 'https://github.com/chewse/pre-commit-mirrors-pydocstyle'
+  #     rev: v2.1.1
+  #     hooks:
+  #         - id: pydocstyle
+  #           language_version: python
+  #           args: [
+  #              '--select=D1',
+  #              '--match=(?!test).*\.py']
+
+  # Add the local pydocstyle hook
+  - repo: local
+    hooks:
+      - id: pydocstyle
+        name: pydocstyle
+        entry: pydocstyle
+        language: python
+        language_version: python3
+        types: [python]
+        args: ['--select=D1', '--match=(?!test).*\.py']
+        additional_dependencies: ['pydocstyle>=6.3.0']
+
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: 'v1.11.2'
+    hooks:
+      - id: mypy
+        args: ['--install-types', '--non-interactive', '--no-strict-optional', '--ignore-missing-imports']
